Implicit Bernstein form

Implicit Bernstein Form

This section of the manual describes the functions to compute the implicit Bernstein form of univariate and multivariate monomials.

Implicit Bernstein Form

Implicit Form

Univariate monomials

Multivariate monomials

Implicit Form

BernsteinExpansions.ImplicitForm — Type.

ImplicitForm{T<:Number}

Type defining an implicit Bernstein form.

Fields

array – the vector of vectors holding the Bernstein coefficients

dim – an integer representing the ambient dimension

source

Univariate monomials

BernsteinExpansions.univariate — Function.

univariate(k::Int64, l::Int64, low::N, high::N)::Vector{N} where {N<:AbstractFloat}

Compute the Bernstein coefficients of a univariate monomial.

Input

k – degree of the given monomial

l – degree of the Bernstein polynomial

low – the lower bound of the interval where the Bernstein coefficients are computed

high – the upper bound of the interval the Bernstein coefficients are computed

Output

A vector with floating point entries containing the Bernstein coefficients.

source

univariate(k::Int64, l::Int64, low::Rational, high::Rational)::Vector{Rational}

Compute exactly the Bernstein coefficients of a univariate monomial.

Input

k – degree of the given monomial

l – degree of the Bernstein polynomial

low – the lower bound of the interval where the Bernstein coefficients are computed

high – the upper bound of the interval the Bernstein coefficients are computed

Output

A vector with rational entries containing the Bernstein coefficients.

source

Multivariate monomials

BernsteinExpansions.multivariate — Function.

multivariate(k::Vector{Int64}, l::Vector{Int64},
                      low::Vector{N}, high::Vector{N})::ImplicitForm{N} where {N<:Number}

Compute the Bernstein coefficients of a multivariate monomial.

Input

k – vector of degrees for each monomial

l – vector of degrees of the Bernstein polynomial for each monomial

low – the lower bounds of the interval where the Bernstein coefficients are computed

high – the upper bounds of the interval the Bernstein coefficients are computed

Output

A Bernstein implicit form holding the Bernstein coefficients.

Examples

julia> m = multivariate([3,2],[3,2],[1.0,2],[2,4.0]);
julia> m.array
2-element Array{Array{Float64,1},1}:
 [1.0, 2.0, 4.0, 8.0]
 [4.0, 8.0, 16.0]

source