Tensorial Bernstein form

Tensorial Bernstein Form

This section of the manual describes the functions to compute the tensorial Bernstein form of univariate and multivariate monomials.

Tensorial Bernstein Form

Tensorial Form

Multivariate monomials

Tensorial Form

BernsteinExpansions.generate_tensor_form — Function.

generate_tensor_form(implicitform::ImplicitForm{T})::Vector{T} where{T<:Number}

Compute the Bernstein expansion in tensorial form given its implicit form.

Input

implicitform – Bernstein expansion in implicit form

Algorithm

This implementation uses Julia's Kronecker product kron function.

source

Multivariate monomials

BernsteinExpansions.tensor_multivariate — Function.

tensor_multivariate(k::Vector{Int64}, l::Vector{Int64},
                      low::Vector{T}, high::Vector{T})::Vector{T} where {T<:Number}

Compute the Bernstein coefficients of a multivariate monomial in the tensor form.

Input

k – vector of degrees for each monomial

l – vector of degrees of the Bernstein polynomial for each monomial

low – the lower bounds of the interval where the Bernstein coefficients are computed

high – the upper bounds of the interval the Bernstein coefficients are computed

Algorithm

This implementation uses Julia's Kronecker product kron function.

source