Types

This section includes the types implemented in BernsteinExpansions.jl.

Types

Bernstein forms

Interface functions

Bernstein forms

BernsteinExpansions.AbstractBernsteinForm — Type

AbstractBernsteinForm{T}

Abstract supertype for all Bernstein forms types.

Notes

This type is parametric in the numeric type T.

julia> using BernsteinExpansions: AbstractBernsteinForm

julia> subtypes(AbstractBernsteinForm)
2-element Vector{Any}:
 FullBernsteinForm
 ImplicitBernsteinForm

BernsteinExpansions.ImplicitBernsteinForm — Type

ImplicitBernsteinForm{T, PT<:AbstractPolynomialLike{T},
                      D, N, TC, NC, SC} <: AbstractBernsteinForm{T}

Type defining an implicit Bernstein form of a polynomial over a hyperrectangular domain.

Fields

pol – polynomial in one or several variables

dom – hyperrectangular set of type IntervalBox representing the domain of the expansion

numvars – number of variables in the polynomial

coeffs – vector of arrays that holds the Bernstein coefficients implicitly

numcoeffs – integer holding the number of Bernstein coefficients that are implicitly defined

degs – vector of degrees of the Bernstein polynomial for each variable

Examples

BernsteinExpansions.FullBernsteinForm — Type

FullBernsteinForm{T, PT<:AbstractPolynomialLike{T},
                  D, N, MC} <: AbstractBernsteinForm{T}

Type defining a full (i.e. explicit) Bernstein form of a polynomial over a hyperrectangular domain.

Fields

pol – polynomial in one or several variables

dom – hyperrectangular set of type IntervalBox representing the domain of the expansion

numvars – number of variables in the polynomial

coeffs – multi-dimensional array that holds the Bernstein coefficients

numcoeffs – integer holding the number of Bernstein coefficients that are implicitly defined

degs – vector of degrees of the Bernstein polynomial for each variables

Examples

Interface functions

BernsteinExpansions.numvars — Function

numvars(::AbstractBernsteinForm)

Return the number of variables of the given Bernstein form.

Input

bf – Bernstein form

Output

The number of variables of the polynomial represented by bf.

BernsteinExpansions.polynomial — Function

polynomial(::AbstractBernsteinForm)

Return the polynomial represented by the given Bernstein form.

Input

bf – Bernstein form

Output

The polynomial associated to bf.

BernsteinExpansions.domain — Function

domain(::AbstractBernsteinForm)

Return the domain over which the given Bernstein form was computed.

Input

bf – Bernstein form

Output

The hyperrectangular domain of bf, of type IntervalOrIntervalBox.

BernsteinExpansions.enclose — Function

enclose(bf::AbstractBernsteinForm)

Compute a range enclosure of the polynomial represented by bf over its own domain, domain(bf).

Input

bf – Bernstein form

Output

An interval (or interval box) that is guaranteed to contain polynomial(bf)(x) for every x in domain(bf) — the minimum and maximum Bernstein coefficients of bf bound the range of the polynomial.

enclose(::AbstractBernsteinForm, ::IntervalOrIntervalBox)

Compute a range enclosure of the polynomial represented by the given Bernstein form.

Input

bf – Bernstein form

X – hyperrectangular domain over which to compute the enclosure

Output

An interval (or interval box) that contains the range of the polynomial over X.

Notes

TODO: not implemented yet.

BernsteinExpansions.getcoeff — Function

getcoeff(::AbstractBernsteinForm, ids::NTuple{D, Int}) where {D}

Return the coefficient of the Bernstein expansion corresponding to the tuple ids.

Input

bf – Bernstein form

ids – tuple of per-variable indices identifying the requested coefficient

Output

The Bernstein coefficient of bf at multi-index ids.