Types

This section includes the types implemented in BernsteinExpansions.jl.

Types

Bernstein forms

Bernstein forms

BernsteinExpansions.AbstractBernsteinForm — Type

AbstractBernsteinForm{T}

Abstract supertype for all Bernstein forms types.

Notes

This type is parametric in the numeric type T.

julia> subtypes(AbstractBernsteinForm)

BernsteinExpansions.ImplicitBernsteinForm — Type

ImplicitBernsteinForm{T, PT<:AbstractPolynomialLike{T},
                      D, N, TC, NC, SC} <: AbstractBernsteinForm{T}

Type defining an implicit Bernstein form of a polynomial over a hyperrectangular domain.

Fields

pol – polynomial in one or several variables

dom – hyperrectangular set of type IntervalBox representing the domain of the expansion

numvars – number of variables in the polynomial

coeffs – vector of arrays that holds the Bernstein coefficients implicitly

numcoeffs – integer holding the number of Bernstein coefficients that are implicitly defined

degs – vector of degrees of the Bernstein polynomial for each variable

Examples